migrate specs to page builder system
